--- src/containers/rspec_reports_bar.jsx
@@ -1,128 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import _ from 'lodash';
-import { BarChart } from '../components';
-import { getProjectRspecReports, getProjectRspecReportsSuccess, getProjectRspecReportsFailure,
-  setProjectRspecReportsPage, resetProjectRspecReports } from '../actions/project_reports_actions';
-import { getColors, setOpacity } from '../helpers/chart_helpers';
-
-const lastReports = (reports, number) => {
-  const keys = _.takeRight(_.keys(reports), number);
-  return _.map(keys, key => reports[key]);
-};
-
-const getDuration = (reports, number = 10) => {
-  if (_.isEmpty(reports)) {
-    return _.times(number, _.constant(0));
-  }
-  return _.map(reports, report => report.attributes.summary.duration);
-};
-
-const getStatus = (reports, number = 10) => {
-  const colors = getColors();
-  if (_.isEmpty(reports)) {
-    return _.times(number, _.constant(colors.grey));
-  }
-  return _.map(reports, (report) => {
-    const failureCount = report.attributes.summary.failure_count;
-    if (failureCount > 0) {
-      return colors.red;
-    }
-    return colors.green;
-  });
-};
-
-const setAllOpacity = (colors, opacity = 1) => _.map(colors, c => setOpacity(c, opacity));
-
-const getChartData = (reports, displayedNumber) => {
-  if (typeof displayedNumber !== 'number') {
-    throw new Error(`Filter ${displayedNumber} is not supported`);
-  }
-  const last = lastReports(reports, displayedNumber);
-  const data = getDuration(last, displayedNumber);
-  const colors = getStatus(last, displayedNumber);
-
-  return {
-    labels: _.times(displayedNumber, _.constant('')),
-    datasets: [
-      {
-        fillColor: setAllOpacity(colors, 0.4),
-        strokeColor: setAllOpacity(colors, 0.7),
-        highlightFill: setAllOpacity(colors, 0.6),
-        highlightStroke: setAllOpacity(colors, 1),
-        data,
-      },
-    ],
-  };
-};
-
-const chartOptions = {
-  barStrokeWidth: 2,
-  barValueSpacing: 3,
-  scaleLineWidth: 3,
-  scaleFontFamily: "'Helvetica Neue'",
-  scaleFontSize: 15,
-  scaleShowVerticalLines: false,
-  responsive: true,
-  maintainAspectRatio: false,
-  tooltipFillColor: 'rgba(255,165,91,0.8)',
-  tooltipTemplate: '<%= value %> seconds',
-};
-
-class RspecReportsBar extends Component {
-  constructor(props) {
-    super(props);
-    this.fetchProjectRspecReports = this.fetchProjectRspecReports.bind(this);
-  }
-
-  componentDidMount() {
-    const { reportsList } = this.props;
-    if (!reportsList || _.isEmpty(reportsList)) {
-      this.fetchProjectRspecReports(this.props.perPage);
-    }
-  }
-
-  componentWillUnmount() {
-    this.props.resetProjectRspecReports();
-  }
-
-  fetchProjectRspecReports(perPage) {
-    const { projectName, xApiKey, dispatch } = this.props;
-    const options = { page: this.props.page, per_page: perPage };
-    getProjectRspecReports(projectName, xApiKey, options)
-      .then((response) => {
-        if (response.status !== 200) {
-          return dispatch(getProjectRspecReportsFailure(response.payload));
-        }
-        dispatch(setProjectRspecReportsPage(response));
-        return dispatch(getProjectRspecReportsSuccess(response));
-      });
-  }
-
-  render() {
-    return (<BarChart
-      reportsCount={this.props.perPage}
-      setFilterCount={this.fetchProjectRspecReports}
-      getChartData={getChartData}
-      options={chartOptions}
-      reports={this.props.reportsList}
-    />);
-  }
-}
-
-const mapDispatchToProps = dispatch => ({
-  getProjectRspecReports: (...args) => dispatch(getProjectRspecReports(...args)),
-  resetProjectRspecReports: () => dispatch(resetProjectRspecReports()),
-  dispatch,
-});
-
-const mapStateToProps = ({ projectReports, users }, ownProps) => ({
-  projectName: ownProps.projectName,
-  page: projectReports.rspecReportsList.page,
-  perPage: projectReports.rspecReportsList.perPage,
-  total: projectReports.rspecReportsList.total,
-  reportsList: projectReports.rspecReportsList.data,
-  xApiKey: users.currentUser.xApiKey,
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(RspecReportsBar);
