+++ src/index.jsx
@@ -16,10 +16,10 @@ import UpdateProject from './containers/update_project';
 import UpdateUser from './containers/update_user';
 import Project from './containers/project';
 import Projects from './containers/projects';
+import ProjectRspecReports from './containers/project_rspec_reports';
 import RequireAuth from './containers/authentication';
 import Reports from './containers/reports';
-import Report from './containers/report';
-import RspecReports from './containers/rspec_reports';
+import RspecReport from './containers/rspec_report';
 import CreateUser from './containers/create_user';
 import User from './containers/user';
 import Users from './containers/users';
@@ -44,11 +44,11 @@ ReactDOM.render(
             <Route path="/users/:id" component={RequireAuth(User)} />
             <Route path="/users" component={RequireAuth(Users)} />
             <Route path="/projects/:name/edit" component={RequireAuth(UpdateProject)} />
-            <Route path="/projects/:name/rspec" component={RequireAuth(RspecReports)} />
+            <Route path="/projects/:name/rspec" component={RequireAuth(ProjectRspecReports)} />
             <Route path="/projects/:name" component={RequireAuth(Project)} />
             <Route path="/projects" component={RequireAuth(Projects)} />
             <Route path="/project/new" component={RequireAuth(CreateProject)} />
-            <Route path="/reports/:id" component={RequireAuth(Report)} />
+            <Route path="/reports/:id" component={RequireAuth(RspecReport)} />
             <Route path="/reports" component={RequireAuth(Reports)} />
             <Route exact path="/" component={Home} />
             <Route path="*" component={NotFound} />
