+++ style/assets/css/gsdk.css
@@ -160,7 +160,8 @@ input[type="file"] > input[type="button"]::-moz-focus-inner {
 .ct-orange.radio.checked .second-icon,
 .ct-orange.checkbox.checked,
 .ct-orange.radio.checked {
-  color: #FF9500; }
+  color: #FF9500;
+  }
 
 .ct-red,
 .ct-red.checkbox.checked .second-icon,
